WotCon tests: split monolithic WotConTests.cs into per-theme partials
Pure code-move; no behavioural changes. 18/18 tests still pass.

Splits the 739-line WotConTests.cs into 8 partials under
tests/CompanionSpecs/WotCon/ so each plan item (Action -> Method,
HasWoTComponent, simulation, optional management members, ...) lands tests in
its own file rather than growing a monolithic class.

Layout:
- WotConTests.cs                          - shell: ctor, NodeId constants, ns helpers, [TestFixture]
- WotConTests.Helpers.cs                  - shared CallAsync / OpenWriteAsync / browse / upload helpers
- WotConTests.Discovery.cs                - namespace registration + management-object discovery (2)
- WotConTests.CreateAsset.cs              - CreateAsset behaviour + browseability + IWoTAssetType (4)
- WotConTests.DeleteAsset.cs              - DeleteAsset happy-path + BadNotFound (2)
- WotConTests.WoTFile.cs                  - File-API round-trip + handle isolation (3)
- WotConTests.CloseAndUpdate.cs           - TD parsing + semantic-validation failures (3)
- WotConTests.PropertyMaterialization.cs  - 6.3.8 Variable mapping (4)
